waitm-637: move to views
